DAOS-2098 control: reorganise package and file naming
Housekeeping changes to keep packages organised.

  - rename app package file with main() to main.go
  - prefix server pkg files with {config,mgmt,storage}
  - restructure utils into flat common pkg
  - move custom log.go into top-level package
  - move proto bindings into common pkg
  - move client/mgmt to client
  - explicitly remove then copy init files in build
    to improve reliability
