+++ src/control/agent/main.go
@@ -1,5 +1,5 @@
 //
-// (C) Copyright 2018 Intel Corporation.
+// (C) Copyright 2018-2019 Intel Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -25,61 +25,53 @@ package main
 
 import (
 	"flag"
-	"fmt"
 	"os"
 	"os/signal"
+	"path/filepath"
 	"syscall"
 
 	"github.com/daos-stack/daos/src/control/drpc"
-	"github.com/daos-stack/daos/src/control/drpc_test/hello"
-	"github.com/daos-stack/daos/src/control/utils/log"
-	"github.com/golang/protobuf/proto"
-	"github.com/pkg/errors"
+	"github.com/daos-stack/daos/src/control/log"
 )
 
 var (
-	unixSocket = flag.String("unix_socket", "./drpc_test.sock", "The path to the unix socket to be used for drpc messages")
-	server     = flag.Bool("server", false, "Start up a dRPC test server. Otherwise, will run in client mode")
+	runtimeDir = flag.String("runtime_dir", "/var/run/daos_agent", "The path to runtime socket directory for daos_agent")
 )
 
 func main() {
+	var err error
+	defer func() {
+		status := 0
+		if err != nil {
+			status = 1
+		}
+		os.Exit(status)
+	}()
+
 	// Set default global logger for application.
 	log.NewDefaultLogger(log.Debug, "", os.Stderr)
 
 	flag.Parse()
 
-	err := errors.New("")
-	if *server {
-		err = runDrpcServer()
-	} else {
-		err = runDrpcClient()
-	}
-	status := 0
-	if err != nil {
-		log.Errorf(err.Error())
-		status = 1
-	}
-	os.Exit(status)
-}
-
-func runDrpcServer() error {
 	// Setup signal handlers so we can block till we get SIGINT or SIGTERM
 	signals := make(chan os.Signal, 1)
 	finish := make(chan bool, 1)
 	signal.Notify(signals, syscall.SIGINT, syscall.SIGTERM)
 
-	drpcServer, err := drpc.NewDomainSocketServer(*unixSocket)
+	sockPath := filepath.Join(*runtimeDir, "agent.sock")
+	drpcServer, err := drpc.NewDomainSocketServer(sockPath)
 	if err != nil {
-		return errors.Wrap(err, "creating socket server")
+		log.Errorf("Unable to create socket server: %v", err)
+		return
 	}
 
-	module := &hello.HelloModule{}
+	module := &SecurityModule{}
 	drpcServer.RegisterRPCModule(module)
 
 	err = drpcServer.Start()
 	if err != nil {
-		return errors.Wrapf(
-			err, "starting socket server on %s", *unixSocket)
+		log.Errorf("Unable to start socket server on %s: %v", sockPath, err)
+		return
 	}
 
 	// Anonymous goroutine to wait on the signals channel and tell the
@@ -92,49 +84,5 @@ func runDrpcServer() error {
 		finish <- true
 	}()
 	<-finish
-
-	return nil
-}
-
-func runDrpcClient() error {
-	client := drpc.NewClientConnection(*unixSocket)
-
-	err := client.Connect()
-	if err != nil {
-		return errors.Wrap(err, "connecting to socket")
-	}
-
-	message := &drpc.Call{
-		Module: int32(hello.Module_HELLO),
-		Method: int32(hello.Function_GREETING),
-	}
-
-	body := &hello.Hello{
-		Name: "Friend",
-	}
-	message.Body, err = proto.Marshal(body)
-	if err != nil {
-		return errors.Wrap(err, "marshalling the Call body")
-	}
-
-	resp, err := client.SendMsg(message)
-	if err != nil {
-		return errors.Wrap(err, "sending message")
-	}
-
-	fmt.Printf("Response:")
-	fmt.Printf("\tSequence: %v", resp.Sequence)
-	fmt.Printf("\tStatus: %v", resp.Status)
-	fmt.Printf("\tBody: %v bytes", len(resp.Body))
-
-	respBody := &hello.HelloResponse{}
-	err = proto.Unmarshal(resp.Body, respBody)
-	if err != nil {
-		return errors.Wrap(err, "unmarshalling HelloResponse")
-	}
-
-	log.Debugf("\tGreeting: %v", respBody.Greeting)
-
-	log.Debugf("Done.")
-	return nil
+	drpcServer.Shutdown()
 }
