+++ src/control/client/mocks.go
@@ -21,33 +21,42 @@
 // portions thereof marked with this legend must also reproduce the markings.
 //
 
-package main
+package mgmtclient
 
 import (
-	"testing"
+	"fmt"
 
-	. "github.com/daos-stack/daos/src/control/utils/test"
-
-	pb "github.com/daos-stack/daos/src/control/proto/mgmt"
+	pb "github.com/daos-stack/daos/src/control/common/proto/mgmt"
 )
 
-func TestGetFeature(t *testing.T) {
-	s := &controlService{}
-
-	mockFeature := MockFeaturePB()
-	fMap := make(FeatureMap)
-	fMap[mockFeature.Fname.Name] = mockFeature
-	s.supportedFeatures = fMap
-
-	feature, err := s.GetFeature(nil, mockFeature.Fname)
-	if err != nil {
-		t.Fatal(err)
+// NewClientFM provides a mock ClientFeatureMap for testing.
+func NewClientFM(features []*pb.Feature, addrs Addresses) ClientFeatureMap {
+	cf := make(ClientFeatureMap)
+	for _, addr := range addrs {
+		fMap := make(FeatureMap)
+		for _, f := range features {
+			fMap[f.Fname.Name] = fmt.Sprintf(
+				"category %s, %s", f.Category.Category, f.Description)
+		}
+		cf[addr] = fMap
 	}
+	return cf
+}
 
-	AssertEqual(t, feature, mockFeature, "")
-
-	feature, err = s.GetFeature(nil, &pb.FeatureName{Name: "non-existent"})
-	if err == nil {
-		t.Fatal(err)
+// NewClientNvme provides a mock ClientNvmeMap for testing.
+func NewClientNvme(ctrlrs NvmeControllers, addrs Addresses) ClientNvmeMap {
+	cMap := make(ClientNvmeMap)
+	for _, addr := range addrs {
+		cMap[addr] = ctrlrs
 	}
+	return cMap
+}
+
+// NewClientScm provides a mock ClientScmMap for testing.
+func NewClientScm(mms ScmModules, addrs Addresses) ClientScmMap {
+	cMap := make(ClientScmMap)
+	for _, addr := range addrs {
+		cMap[addr] = mms
+	}
+	return cMap
 }
