+++ src/control/client/scm.go
@@ -27,7 +27,7 @@ import (
 	"io"
 	"time"
 
-	pb "github.com/daos-stack/daos/src/control/proto/mgmt"
+	pb "github.com/daos-stack/daos/src/control/common/proto/mgmt"
 
 	"golang.org/x/net/context"
 )
