+++ src/control/dmg/main.go
@@ -28,8 +28,8 @@ import (
 	"os"
 	"strings"
 
-	"github.com/daos-stack/daos/src/control/client/mgmt"
-	"github.com/daos-stack/daos/src/control/utils/log"
+	"github.com/daos-stack/daos/src/control/client"
+	"github.com/daos-stack/daos/src/control/log"
 
 	flags "github.com/jessevdk/go-flags"
 	"github.com/pkg/errors"
