+++ src/control/run_go_tests.sh
@@ -11,7 +11,7 @@ echo "GOPATH $GOPATH"
 repopath=github.com/daos-stack/daos
 
 # Lint source then run Go tests for each package
-for d in client/mgmt security server dmg drpc; do
+for d in client security server dmg drpc; do
     echo "testing $d"
     pushd "$GOPATH/src/$repopath/src/control/$d"
     # todo: provide a sensible way of linting and returning review comments
