+++ src/control/security/auth_sys_test.go
@@ -26,9 +26,9 @@ package security_test
 import (
 	"testing"
 
+	. "github.com/daos-stack/daos/src/control/common"
 	"github.com/daos-stack/daos/src/control/security"
 	pb "github.com/daos-stack/daos/src/control/security/proto"
-	. "github.com/daos-stack/daos/src/control/utils/test"
 
 	"github.com/golang/protobuf/proto"
 )
