+++ src/control/server/storage_nvme_test.go
@@ -28,8 +28,8 @@ import (
 	"strings"
 	"testing"
 
-	pb "github.com/daos-stack/daos/src/control/proto/mgmt"
-	. "github.com/daos-stack/daos/src/control/utils/test"
+	. "github.com/daos-stack/daos/src/control/common"
+	pb "github.com/daos-stack/daos/src/control/common/proto/mgmt"
 	. "github.com/daos-stack/go-spdk/spdk"
 )
